+++ hivemind/utils/__init__.py
@@ -2,7 +2,7 @@ from hivemind.utils.connection import *
 from hivemind.utils.nested import *
 from hivemind.utils.tensor_descr import *
 from hivemind.utils.serializer import *
-from hivemind.utils.shared_future import *
+from hivemind.utils.mpfuture import *
 from hivemind.utils.threading import *
 from hivemind.utils.autograd import *
 from hivemind.utils.grpc import *
